+++ src/Kephas.Services.Abstractions/release-notes.md
@@ -4,5 +4,4 @@
 * Documentation and samples: https://github.com/kephas-software/kephas/wiki and https://github.com/kephas-software/kephas/tree/master/Samples.
 
 # 12.0.0
- 
-* Added  ```release-notes.md```.
+* Initial branch from `Kephas.Abstractions`.